move spack.llnl.util.lock into spack.util.lock
