+++ lib/spack/spack/test/util/lock.py
@@ -6,7 +6,7 @@
 
 Run with pytest::
 
-    pytest lib/spack/spack/test/llnl/util/lock.py
+    pytest lib/spack/spack/test/util/lock.py
 
 You can use this to test whether your shared filesystem properly supports
 POSIX reader-writer locking with byte ranges through fcntl.
@@ -43,9 +43,8 @@ from contextlib import contextmanager
 from multiprocessing import Barrier, Process, Queue
 
 import pytest
-
-import spack.llnl.util.lock as lk
 from spack.util.filesystem import getuid, touch, working_dir
+import spack.util.lock as lk
 
 if sys.platform != "win32":
     import fcntl
